refactor: support long token
